requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
google-generativeai
streamlit